requirements.txt
mysql-connector-python~=9.0.0
openai~=1.35.7
pandas~=2.2.2
numpy~=2.0.0
reportlab~=4.2.2
matplotlib~=3.9.0